temp: install python-ldap via pip until 3.12 package
The distro's Python-ldap package is not released for 3.12 and it
must be installed via pip. However that requires gcc. So to speed
up installation, the python-ldap package is installed in this image
so the consumers do not need to install gcc and do the necessary
clean up afterwards.

This commit can be reverted (not git revert please) once the
python3.12-ldap package is ready (or python3.12 becomes
the default Python version).

JIRA: SYSDEV-2145